# back/docs/swagger.yaml
definitions:
  types.ComplaintPayload:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  types.FormularioDados:
    properties:
      atuacao:
        type: string
      cidade:
        type: string
      email:
        type: string
      empresa:
        type: string
      name:
        type: string
      porte:
        type: string
      servicos:
        items:
          type: string
        type: array
      setor:
        type: string
      telefone:
        type: string
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Healthcheck endpoint
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Healthcheck
      tags:
      - ops
  /v1/complaints/email:
    post:
      consumes:
      - application/json
      description: Recebe uma reclamação e envia email de confirmação para o cliente
        e notificação para a equipa de suporte
      parameters:
      - description: Complaint Info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.ComplaintPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ComplaintPayload'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Envia reclamação do cliente
      tags:
      - complaints
  /v1/enterprises/email:
    post:
      consumes:
      - application/json
      description: Recebe dados de uma empresa, envia um email de boas-vindas ao cliente
        e uma notificação para a equipa de vendas
      parameters:
      - description: Enterprise Info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.FormularioDados'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.FormularioDados'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Envia email de confirmação e notificação
      tags:
      - enterprises
swagger: "2.0"